+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 ## Release Notes
 
 - 0.5.1 *Unreleased*
-    - Fix issue with command line args separated by `=` not being parsed correctly.
+    - Fix issue with command line args separated by `=` not being parsed correctly, and improve handling of `-m=a=b` args.
     - Prevent a potential crash if a `nil` `isDirectory` value is force-unwrapped.
     - Correctly `nil` references to `DispatchSourceSignal` instances after cancellation.
     - Add a default error message on `runProcess(()` internal `Process.run()` immediate fails.
     - Abstract away `NSString` functions in `Path.swift` with a `String` extension.
+    - Add `Cli` module tests.
 - 0.5.0 *13 May 2026*
     - Add `Stdio.settings` struct for passing rendering settings.
     - Add emoji support to output as alternative to text.